Handle prime numbers.
Prime numbers suck (in this situation) because they only allow two
possible tile arrangements, and they both usually suck, for common
window sizes. (E.g. 1 row of 13 columns or 13 rows of 1 column.)

This tests for primality (only one set of factors) and if it's prime and
not 2 (you're screwed with 2) it finds the best size again for n + 1
tiles. If that's better (it almost always is) it goes with that.
Essentially we pad a blank square. We could do something fun in that
square if we wanted to.

Most primes are surrounded by more factorable numbers. It works in the
common cases.